[Sessions] _endSession .timestamp equals to .ended_at
Session's `session.ended_at` value already had the correct time (when
the app gone to background). However, datapoint's `timestamp` and
`event.date` will have a date in the future, which is the date when
the data point is tracked. To avoid this, we're setting all the three
dates with the same value.

Signed-off-by: Miguel M. Almeida <miguelcma@gmail.com>